[#163056089] Fix button styling on various asset pages
