+++ src/__test__/components/AssetCategoriesComponent.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import expect from 'expect';
 
-import { AssetCategoriesComponent } from '../../components/AssetCategoriesComponent';
+import { AssetCategoriesComponent } from '../../components/Category/AssetCategoriesComponent';
 
 import assetCategories from '../../_mock/assetCategories';
