+++ src/__test__/components/AssetsSubCategoriesComponent.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import expect from 'expect';
 
-import { AssetSubCategoriesComponent } from '../../components/AssetsSubCategoriesComponent';
+import { AssetSubCategoriesComponent } from '../../components/SubCategory/AssetsSubCategoriesComponent';
 
 import assetSubCategories from '../../_mock/subcategories';
