+++ src/_components/RoutesComponent.jsx
@@ -4,17 +4,17 @@ import jwt from 'jsonwebtoken';
 import { get } from 'lodash';
 
 import Authenticate from './AccessControl/AuthenticateComponent';
-import AssetTypes from '../components/AssetTypesComponent';
+import AssetTypes from '../components/AssetTypes/AssetTypesComponent';
 import Assets from '../components/Assets/AssetsTabComponent';
 import AssetModels from '../components/AssetModels/AssetModelsComponent';
 import LoginComponent from '../components/LoginComponent';
 import Dashboard from '../_components/Dashboard/DashboardContainer';
 import UserFeedback from '../components/User/UserFeedbackComponent';
 import PageNotFound from '../components/PageNotFoundComponent';
-import AssetSubCategories from '../components/AssetsSubCategoriesComponent';
+import AssetSubCategories from '../components/SubCategory/AssetsSubCategoriesComponent';
 import Allocations from '../components/AllocationsComponent';
 import IncidenceReports from '../components/IncidenceReportsComponent';
-import AssetCategories from '../components/AssetCategoriesComponent';
+import AssetCategories from '../components/Category/AssetCategoriesComponent';
 import AssetDetail from './AssetDetails/AssetDetailsContainer';
 import AssetConditions from '../components/AssetCondition/AssetConditionsComponent';
 import AssetMakes from '../components/AssetMake/AssetMakeComponent';
