+++ src/components/AssetModels/AssetModelsComponent.jsx
@@ -12,10 +12,11 @@ import LoaderComponent from '../../components/LoaderComponent';
 import formatDate from '../../_utils/dateFormatter';
 import ModalComponent from '../common/ModalComponent';
 import ModelNumberContainer from '../../_components/ModelNumber/ModelNumberContainer';
+import ItemsNotFoundComponent from '../common/ItemsNotFoundComponent';
 
 import { loadAssetModels } from '../../_actions/assetModels.action';
+
 import '../../_css/AssetsComponent.css';
-import ItemsNotFoundComponent from '../common/ItemsNotFoundComponent';
 
 export class AssetModelsComponent extends React.Component {
   state = {
@@ -67,8 +68,7 @@ export class AssetModelsComponent extends React.Component {
               <ModalComponent
                 trigger={
                   <Button
-                    className="add-asset"
-                    size="medium"
+                    className="filter-button"
                   >
                     ADD MODEL NUMBER
                   </Button>
